Reorganize methods to better files.
