+++ safe/definitionsv4/default_values.py
@@ -3,6 +3,11 @@
 
 from safe.utilities.i18n import tr
 
+__copyright__ = "Copyright 2016, The InaSAFE Project"
+__license__ = "GPL version 3"
+__email__ = "info@inasafe.org"
+__revision__ = '$Format:%H$'
+
 female_ratio_default_value = {
     'key': 'female_ratio_default_value',
     'name': tr('Female Ratio Default Value'),
