--- safe/definitionsv4/settings.py
@@ -1,3 +1,0 @@
-default_values = {
-    'feature_rate_field': 0.1,
-}
