+++ safe/gui/tools/wizard/step_kw47_default_inasafe_fields.py
@@ -28,8 +28,7 @@ from safe.definitionsv4.layer_purposes import (layer_purpose_aggregation)
 from safe.definitionsv4.utilities import get_fields, get_compulsory_fields
 from safe.definitionsv4.layer_geometry import layer_geometry_raster
 from safe.definitionsv4.constants import no_field
-from safe.gui.tools.wizard.wizard_utils import get_defaults
-from safe.utilities.i18n import tr
+from safe.gui.tools.wizard.wizard_utils import get_inasafe_default_value_fields
 
 from safe.gui.tools.wizard.wizard_step import (
     WizardStep, get_wizard_step_ui_class)
@@ -152,9 +151,9 @@ class StepKwDefaultInaSAFEFields(WizardStep, FORM_CLASS):
             parameter.element_type = unicode
             parameter.options_list = option_list
             parameter.value = no_field
-            parameter.default_labels = get_defaults(
+            parameter.default_labels = get_inasafe_default_value_fields(
                 self.parent.setting, inasafe_field['key'])[0]
-            parameter.default_values = get_defaults(
+            parameter.default_values = get_inasafe_default_value_fields(
                 self.parent.setting, inasafe_field['key'])[1]
             # Check if there is already value in the metadata.
             if existing_inasafe_field:
